pyproject.toml
[tool.poetry]
name = "torch-mlops"
version = "0.1.0"
description = "Torch ML"
authors = ["datapsycho <mr.data.psycho@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.1"
PyYAML = "^6.0"
scikit-learn = "^1.1.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
